+++ src/RecaptchaRegistry.php
@@ -10,6 +10,7 @@ use Yiisoft\Translator\TranslatorInterface;
 
 final class RecaptchaRegistry
 {
+    private static ?RecaptchaV3Badge $badge = null;
     private static ?RecaptchaClient $client = null;
 
     /** @var array<string, string>|null */
@@ -17,7 +18,15 @@ final class RecaptchaRegistry
     private static ?string $containerTag = null;
     private static ?bool $containerUseContainer = null;
     private static ?RequestProviderInterface $requestProvider = null;
+    private static ?RecaptchaV2Size $sizeV2 = null;
+    private static ?RecaptchaV2Theme $themeV2 = null;
     private static ?TranslatorInterface $translator = null;
+    private static ?RecaptchaV2Type $typeV2 = null;
+
+    public static function badge(): ?RecaptchaV3Badge
+    {
+        return self::$badge;
+    }
 
     public static function client(): ?RecaptchaClient
     {
@@ -65,6 +74,10 @@ final class RecaptchaRegistry
         self::$containerUseContainer = null;
         self::$containerTag = null;
         self::$containerAttributes = null;
+        self::$badge = null;
+        self::$themeV2 = null;
+        self::$sizeV2 = null;
+        self::$typeV2 = null;
     }
 
     public static function resolveClientIp(?RequestProviderInterface $provider): ?string
@@ -90,6 +103,11 @@ final class RecaptchaRegistry
         }
     }
 
+    public static function setBadgeDefault(?RecaptchaV3Badge $badge): void
+    {
+        self::$badge = $badge;
+    }
+
     /**
      * @param array{useContainer?: bool, tag?: string, attributes?: array<string, string>} $config
      */
@@ -100,8 +118,33 @@ final class RecaptchaRegistry
         self::$containerAttributes = $config['attributes'] ?? null;
     }
 
+    public static function setV2Defaults(
+        ?RecaptchaV2Theme $theme = null,
+        ?RecaptchaV2Size $size = null,
+        ?RecaptchaV2Type $type = null,
+    ): void {
+        self::$themeV2 = $theme;
+        self::$sizeV2 = $size;
+        self::$typeV2 = $type;
+    }
+
+    public static function sizeV2(): ?RecaptchaV2Size
+    {
+        return self::$sizeV2;
+    }
+
+    public static function themeV2(): ?RecaptchaV2Theme
+    {
+        return self::$themeV2;
+    }
+
     public static function translator(): ?TranslatorInterface
     {
         return self::$translator;
+    }
+
+    public static function typeV2(): ?RecaptchaV2Type
+    {
+        return self::$typeV2;
     }
 }